+++ src/monolithic/gst/elements/gvagenai/backends/openvino-genai/openvino-genai_pipeline.cpp
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: MIT
  ******************************************************************************/
 
-#include "openvino_pipeline.hpp"
+#include "openvino-genai_pipeline.hpp"
 
 #include <cmath>
 #include <cstring>
